requirements.txt
aiosqlite~=0.20.0
discord.py~=2.4.0